+++ Lesson_12/Home_work/general_motors.py
