+++ Lesson_12/Home_work/gm_gmc.py
@@ -1,23 +1,20 @@
-from abc import ABC
-
-from vag_class import Vag
+from general_motors import Gm
 
 
-class Vw(Vag, ABC):
+class Gmc(Gm):
     def __init__(self):
         super().__init__()
-        self.brand = 'Vw'
-        self.city_office = 'Wolfsburg'
-        self.motto = 'Das auto'
+        self.brand = 'GMC'
+        self.city_office = 'Detroit'
+        self.motto = 'Everybody in'
 
     def work_long_without_crash(self):
         print(f'{self.brand} has a reliable engine and choir')
 
     def refuel(self):
-        self.work_long_without_crash()
-        print('But I need fuel')
+        print('Please buy more petrol. I like 92!!!')
 
-    def hard_repair(self):
+    def simple_repair(self):
         print(f'Only some services can repair your car - {self.brand}')
 
     def ride(self):
@@ -26,6 +23,8 @@ class Vw(Vag, ABC):
     def say_motto(self):
         print(self.motto)
 
+    def electricity_problem(self):
+        print(f'{self.brand} has troubles wit electricity')
+
     def info_brand(self):
-        self.say_motto()
         print(f'Office of {self.brand} in {self.city_office}')
