+++ Lesson_12/Home_work/gm_opel.py
@@ -1,4 +1,4 @@
-from gm_class import Gm
+from general_motors import Gm
 
 
 class Opel(Gm):
