+++ Lesson_12/Home_work/toyota_co.py
