+++ Lesson_12/Home_work/toyotaco_lexus.py
@@ -1,4 +1,4 @@
-from toyota_co_class import ToyotaCo
+from toyota_co import ToyotaCo
 
 
 class Lexus(ToyotaCo):
