+++ Lesson_12/Home_work/vag.py
